+++ changelog.md
@@ -1,2 +1,7 @@
+### Additions
+- Added a sound to Otherworldly Energy infusing
+- Added some particles when an item is infused by Otherworldly Energy
+
 ### Changes
 - Shatter Coils will no longer attack tamed animals
+- Otherworldly Energy now applies a pickup delay to the result